Splits PBCore models into modules
