+++ .gitignore
@@ -4,3 +4,4 @@ __pycache__
 
 batch3
 pbcore-json
+pbcore-xml
