+++ pbcore/models/assets.py
@@ -1,50 +1,15 @@
 from typing import List, Optional
 
-from pydantic import BaseModel, Field, RootModel
-
-# Base classes for PBCore elements
-
-
-class PBCoreBaseModel(BaseModel):
-    """Base class for all PBCore elements."""
-
-
-class PBCoreTextElement(PBCoreBaseModel):
-    """Base class for a required text elements in PBCore."""
-
-    text: str
-
-
-class PBCoreBaseAttributes(PBCoreBaseModel):
-    """Base class for attributes in PBCore."""
-
-    source: Optional[str] = None
-    ref: Optional[str] = None
-    version: Optional[str] = None
-    annotation: Optional[str] = None
-
-
-class PBCoreAttributesDateType(PBCoreBaseModel):
-    dateType: Optional[str] = None
-
-
-class PBCoreAttributesTime(PBCoreBaseModel):
-    startTime: Optional[str] = None
-    endTime: Optional[str] = None
-    timeAnnotation: Optional[str] = None
-
-
-class PBCoreAttributesAffiliation(PBCoreBaseModel):
-    affiliation: Optional[str] = None
-    affiliationSource: Optional[str] = None
-    affiliationRef: Optional[str] = None
-    affiliationVersion: Optional[str] = None
-    affiliationAnnotation: Optional[str] = None
-
-
-class PBCoreElement(PBCoreTextElement, PBCoreBaseAttributes):
-    """Base class for PBCore elements with required text and optional attributes."""
-
+from pydantic import Field
+from pbcore.models.base import (
+    PBCoreElement,
+    PBCoreAttributesDateType,
+    PBCoreAttributesTime,
+    PBCoreAttributesAffiliation,
+    PBCoreBaseModel,
+    PBCoreTextElement,
+    PBCoreBaseAttributes,
+)
 
 # PBCore Assets
 
@@ -253,99 +218,23 @@ class PBCoreExtension(PBCoreBaseModel):
     extensionEmbedded: Optional[ExtensionEmbedded] = None
 
 
-class InstantiationIdentifier(PBCoreBaseModel):
-    source: str
-    text: str
-
-
-class InstantiationPhysical(PBCoreBaseModel):
-    text: str
-
-
-class InstantiationLocation(PBCoreBaseModel):
-    text: str
-
-
-class InstantiationMediaType(PBCoreBaseModel):
-    text: str
-
-
-class InstantiationGeneration(PBCoreBaseModel):
-    text: str
-
-
-class InstantiationDuration(PBCoreBaseModel):
-    text: str
-
-
-class InstantiationTracks(PBCoreBaseModel):
-    text: str
-
-
-class InstantiationChannelConfiguration(PBCoreBaseModel):
-    text: str
-
-
-class InstantiationAlternativeModes(PBCoreBaseModel):
-    text: str
-
-
-class EssenceTrackType(PBCoreBaseModel):
-    text: str
-
-
-class EssenceTrackAspectRatio(PBCoreBaseModel):
-    text: str
-
-
-class InstantiationEssenceTrack(PBCoreBaseModel):
-    essenceTrackType: EssenceTrackType
-    essenceTrackAspectRatio: Optional[EssenceTrackAspectRatio] = None
-
-
-class InstantiationAnnotation(PBCoreBaseModel):
-    annotationType: Optional[str] = None
-    text: str
-
-
-class PBCoreInstantiation(PBCoreBaseModel):
-    instantiationIdentifier: List[InstantiationIdentifier]
-    instantiationPhysical: Optional[InstantiationPhysical] = None
-    instantiationLocation: InstantiationLocation
-    instantiationMediaType: Optional[InstantiationMediaType] = None
-    instantiationGenerations: Optional[List[InstantiationGeneration]] = None
-    instantiationDuration: Optional[InstantiationDuration] = None
-    instantiationTracks: Optional[InstantiationTracks] = None
-    instantiationChannelConfiguration: Optional[InstantiationChannelConfiguration] = (
-        None
-    )
-    instantiationAlternativeModes: Optional[InstantiationAlternativeModes] = None
-    instantiationEssenceTrack: Optional[List[InstantiationEssenceTrack]] = None
-    instantiationAnnotation: Optional[List[InstantiationAnnotation]] = None
-
-
-class PBCoreDescriptionDocument(PBCoreBaseModel):
-    xsi_schemaLocation: str = Field(..., alias='xsi:schemaLocation')
-    pbcoreAssetType: Optional[List[PBCoreAssetType]] = None
-    pbcoreAssetDate: Optional[List[PBCoreAssetDate]] = Field(None, min_items=1)
-    pbcoreIdentifier: List[PBCoreIdentifier] = Field(..., min_items=1)
-    pbcoreTitle: List[PBCoreTitle] = Field(..., min_items=1)
-    pbcoreSubject: Optional[List[PBCoreSubject]] = None
-    pbcoreDescription: List[PBCoreDescription] = Field(..., min_items=1)
-    pbcoreGenre: Optional[List[PBCoreGenre]] = None
-    pbcoreRelation: Optional[List[PBCoreRelation]] = None
-    pbcoreCoverage: Optional[List[PBCoreCoverage]] = None
-    pbcoreAudienceLevel: Optional[List[PBCoreAudienceLevel]] = None
-    pbcoreAudienceRating: Optional[List[PBCoreAudienceRating]] = None
-    pbcoreCreator: Optional[List[PBCoreCreator]] = None
-    pbcoreContributor: Optional[List[PBCoreContributor]] = None
-    pbcorePublisher: Optional[List[PBCorePublisher]] = None
-    pbcoreRightsSummary: Optional[List[PBCoreRightsSummary]] = None
-    pbcoreInstantiation: Optional[List[PBCoreInstantiation]] = None
-    pbcoreAnnotation: Optional[List[PBCoreAnnotation]] = None
-    pbcorePart: Optional[List[PBCorePart]] = None
-    pbcoreExtension: Optional[List[PBCoreExtension]] = None
-
-
-class PBCore(PBCoreBaseModel):
-    pbcoreDescriptionDocument: PBCoreDescriptionDocument
+__all__ = [
+    "PBCoreAssetType",
+    "PBCoreAssetDate",
+    "PBCoreIdentifier",
+    "PBCoreTitle",
+    "PBCoreSubject",
+    "PBCoreDescription",
+    "PBCoreGenre",
+    "PBCoreRelation",
+    "PBCoreCoverage",
+    "PBCoreAudienceLevel",
+    "PBCoreAudienceRating",
+    "PBCoreCreator",
+    "PBCoreContributor",
+    "PBCorePublisher",
+    "PBCoreRightsSummary",
+    "PBCorePart",
+    "PBCoreAnnotation",
+    "PBCoreExtension",
+]
